- Design changed to adapt the BottomNavigationView from support library
We are using LockableViewPager( A viewpager without swipe touch support) with BottomNavigationView to handle the fragments as the viewpager will take care of saving state of all the fragment.
There is an alternative approach chosen initially to make separate Framelayout and add fragment to them for each BottomNavigation tab but then we have to manually set visibility of other fragment containe framelayout's visibility to GONE and this may creates the issue of stackoverflowexception thought it wasnt happened but its too heavy implementation i guess. not still sure but this viewpager is pretty straightforward compare to the this framelayout holder for each fragment!

Also this approach is used by Freelancer app too!