- BottomNavigationView replaced with tablayout as tablayout can be setupWithViewpager perfectly as they ahve this method. Also BottomNavigationView has some problem in making selection when viewpager scrolled to new page and the hack of RichBottomNavigationView works but still the animation is so weird in dynamic selection.  The issues with BottomNavigation in summary :- 1) dynamic selection animation problem 2) parallax effect of viewpager emulate on its background is very tricky  and we have just cropped the bitmap to do so but not the good workaround
Whereas the Tablayout offers good alternative as works with viewpager so the parallax effect achieved with no effort. also switched to text based strip as text is more user appealing than the icon. User has to guess about icons but when we use text i think user will be happy with text.