Replace library & capital popup letter
The jaredrummler's fast-scroll is fork of the timusus's library.
The issue was sometimes it creates issue where adapter is set but due
to no null-check in their code it was throwing NPE. Replace
jaredrummler's fast-scroll by timusus's fast-scroll library.

But the original library from timusus is constantly updated and has
no adapter NPE issue.