Get rid of obsolete layout params
The given layout_param is not defined for the given layout, meaning it
has no effect. This usually happens when you change the parent layout or
move view code around without updating the layout params. This will cause
useless attribute processing at runtime, and is misleading for others
reading the layout so the parameter should be removed.