Check NowPlayingList is not empty before playing song
When user removes the app from "recent apps" and then if user dial
call, and app automatically crashes. The reason is very weird.

If user dont remove the app from recent and dial the call then the
function of call reciever ( pause when call is running and resume
when disconnected) works fine. But If user removes the app from
recent and then dial the call then NowPlayingList automatically
becomes empty!

Due to this facts, isEmpty() check is necessary wherever we use
currentMusic() method.