Use ActionItem.OnClickListener & Remove  QuickAction/Listener classes
MemberQuickAction,MusicQuickAction & PlaylistQuickAction and their
listeners are just duplicating things over and over and very
inefficient. Get rid of them and add ActionItem.OnClickListener as
it avoids those unnecessary (ABC)QuickActionListener interfaces too!

Fixes the color update probem of IconDrawable in QuickActionWindow
(As now it doesnot create the QuickActionWindow in adapter's
constructor but create it on demand)