Fix: Check if view attached before calling onPaletteOrThemeEvent()
According to firebase crashes, it seems that presenter sometimes
presenter is not prepared yet and takeView() method not called on it
and so it throws NPE on view().refresh()! This happens due to the
problem of subscribeEvents(). We should subscribeEvents() only after
presenter prepared properly, but as not sure when to unsubscribe..

For now its temporary fix to check if presenter has view in fragment
as we dont want this null check boilerplate in presenter for every
method!