Introduce Group interface with sorting behaviour
Collections.sort() was heavily used everywhere in codebase and its
simply code duplication. Get rid of it with MusicGroup which
encapsulate all the logic to sort music.

Move the SortEvent.SortBY annotation to its separate file as Sort
mechanism is not coupled only with sorting event but can be used
generally everywhere.