Handle RequiredPermissions in MainActivity to avoid crashes
We were getting issues when user allows all permission but then after
sometime they go to app permission and disable the permission. As
splash activity handle the Required runtime permission so if user
then launch the app it can work perfect without issue. But here if
app is already open and MainActivity is on top stack then after
disabling permission when user comeback to app, simply app crashes on
user's face!

Just add check if permission not granted in main
 screen then redirect to splash screen again.