Update song info on each MusicEvent if currentPlayingSong is changed
It was only updating song info if MusicPlayerEvent is NEXT or PREVIOUS
which works perfect but since OutOfMemoryError we were getting due to
load of album and lock screen image on each MusicPlayerEvent, we have
debounced MusicEvent stream. Due to this fact we were not getting NEXT
& PREVIOUS event as they are immediately followed by PLAY event.

To avoid this problem where song being played is different than the
song info displayed, we have to always make sure by checking that
currentPlayingSong is not equals to MusicEvent then display the new
song info.