Fix bug: where last music plays on click of new music
Finally this issue fixed. It was here since very long but due to
unpredicatability of its happening it couldnt be solved but finally
noticed that it was happening everytime when user pause the player
and then click on new song would always resumes the last song.

Approach is that whenever user jump to any song in NowPlayingList,
then Music Player should always call setDataSource()