Dont put last album bitmap in GlideBitmapPool
It was crashing the app after few NEXT/PREVIOUS event with "Can't
blur bitmap". The reason must be the album bitmap. As album bitmap
already we are getting from Glide call so its memory already managed by
Glide itself so i think it should not be put in this GlideBitmapPool to
unload its memory twice. Whereas the lockscreen bitmap is not Managed
by Glide so we have to manage it by putting it in GlideBitmapPool.