Add class to provide unique player actions based on package name
This is necessary as right now when both free and paid app installed
on device is creating issues with broadcast system. Both free and
paid app dont handle action as they both having the same action string.
It is must to have unique action string for both app to perform actions.