Use Music interface only & Extend the Parcelable interface
App heavily was relying SqlMusic which is very wrong. So lets use Music
interface only. But Compiler throws this error:
"List<Music> cannot be converted to ArrayList<? extends Parcelable>"
Due to this only problem we had used SqlMusic type everywhere so lets
accept that Parcelable in Music Interface to let it compile!